// lib/types/threejs.type.ts
export type TVType = GLTF & {
  nodes: {
    Object_25: THREE.Mesh;
    Object_13: THREE.Mesh;
    Object_4: THREE.Mesh;
    Object_22: THREE.Mesh;
    Object_28: THREE.Mesh;
    Object_16: THREE.Mesh;
    Object_19: THREE.Mesh;
    Object_7: THREE.Mesh;
    Object_31: THREE.Mesh;
    Object_10: THREE.Mesh;
    Object_31001: THREE.Mesh;
  };
  materials: {
    ["02_-_Default"]: THREE.MeshStandardMaterial;
    ["03_-_Default"]: THREE.MeshStandardMaterial;
    ["04_-_Default"]: THREE.MeshStandardMaterial;
    ["05_-_Default"]: THREE.MeshStandardMaterial;
    main_glass: THREE.MeshPhysicalMaterial;
    ["screen_main.001"]: THREE.MeshStandardMaterial;
  };
};